\name{fil.fas}
\alias{fil.fas}
\docType{data}
\title{
Fasta data example.
}
\description{
Internal Fasta format example
}
\usage{data(fil.fas)}
\format{
  Class 'fasta'
  with elements containing ">"indicating sequence names and the DNA sequences in subsequent elements.  
}
\source{
From Qiong Ding
}
\references{
None.
}
\examples{
data(fil.fas)
fil.fas
}
\keyword{datasets}
