\name{fil.phy}
\alias{fil.phy}
\docType{data}
\title{
Internal phylip format data
}
\description{
Provide an example of phylip data format.
}
\usage{data(fil.phy)}
\format{
 class 'phy'.
 With each line stored as one element in a character vector.
}
\details{
To be modified later.
}
\source{
Provided by Qiong Ding.
}
\references{
None.
}
\examples{
data(fil.phy)
fil.phy
}
\keyword{datasets}
